Iterators: Break up code into several files.
